+++ demos/employee-vci/src/logic/employeeCredential.ts
@@ -2,7 +2,7 @@ import * as jose from "jose";
 
 import store from "../store.js";
 import keyStore from "../../../../src/store/keyStore.js";
-import { issueFlatCredential } from "../../../../src/sd-jwt/issuer.js";
+import { issueFlatCredential } from "../../../../src/credentials/sd-jwt/issuer.js";
 import { ErrorPayload, Result } from "../../../../src/types";
 
 const issueEmployeeCredential = async (
