+++ src/credentials/sd-jwt/types.ts
